requirements.txt
python-telegram-bot==20.5
aiosqlite==0.18.0
nest-asyncio==1.5.6
httpx~=0.24.1